+++ management/model_sign_on_policy_action_id_first_all_of.go
@@ -16,8 +16,6 @@ import (
 
 // SignOnPolicyActionIDFirstAllOf struct for SignOnPolicyActionIDFirstAllOf
 type SignOnPolicyActionIDFirstAllOf struct {
-	// A boolean that specifies whether users must confirm data returned from an identity provider prior to registration. Users can modify the data and omit non-required attributes. Modified attributes are added to the user's profile during account creation. This is an optional property. If omitted, the default value is set to false.
-	ConfirmIdentityProviderAttributes *bool `json:"confirmIdentityProviderAttributes,omitempty"`
 	// The list of IDP discovery rules that are evaluated in order when no user is associated with the user identifier. The maximum number of rules is 100. The condition on which this identity provider is used to authenticate the user is expressed using the PingOne policy condition language
 	DiscoveryRules []SignOnPolicyActionIDFirstAllOfDiscoveryRules `json:"discoveryRules,omitempty"`
 	// A boolean that if set to true and if the user's account is locked (the account.canAuthenticate attribute is set to false), then social sign on with an external identity provider is prevented.
@@ -34,8 +32,6 @@ type SignOnPolicyActionIDFirstAllOf struct {
 // will change when the set of required properties is changed
 func NewSignOnPolicyActionIDFirstAllOf() *SignOnPolicyActionIDFirstAllOf {
 	this := SignOnPolicyActionIDFirstAllOf{}
-	var confirmIdentityProviderAttributes bool = false
-	this.ConfirmIdentityProviderAttributes = &confirmIdentityProviderAttributes
 	return &this
 }
 
@@ -44,41 +40,7 @@ func NewSignOnPolicyActionIDFirstAllOf() *SignOnPolicyActionIDFirstAllOf {
 // but it doesn't guarantee that properties required by API are set
 func NewSignOnPolicyActionIDFirstAllOfWithDefaults() *SignOnPolicyActionIDFirstAllOf {
 	this := SignOnPolicyActionIDFirstAllOf{}
-	var confirmIdentityProviderAttributes bool = false
-	this.ConfirmIdentityProviderAttributes = &confirmIdentityProviderAttributes
 	return &this
-}
-
-// GetConfirmIdentityProviderAttributes returns the ConfirmIdentityProviderAttributes field value if set, zero value otherwise.
-func (o *SignOnPolicyActionIDFirstAllOf) GetConfirmIdentityProviderAttributes() bool {
-	if o == nil || o.ConfirmIdentityProviderAttributes == nil {
-		var ret bool
-		return ret
-	}
-	return *o.ConfirmIdentityProviderAttributes
-}
-
-// GetConfirmIdentityProviderAttributesOk returns a tuple with the ConfirmIdentityProviderAttributes field value if set, nil otherwise
-// and a boolean to check if the value has been set.
-func (o *SignOnPolicyActionIDFirstAllOf) GetConfirmIdentityProviderAttributesOk() (*bool, bool) {
-	if o == nil || o.ConfirmIdentityProviderAttributes == nil {
-		return nil, false
-	}
-	return o.ConfirmIdentityProviderAttributes, true
-}
-
-// HasConfirmIdentityProviderAttributes returns a boolean if a field has been set.
-func (o *SignOnPolicyActionIDFirstAllOf) HasConfirmIdentityProviderAttributes() bool {
-	if o != nil && o.ConfirmIdentityProviderAttributes != nil {
-		return true
-	}
-
-	return false
-}
-
-// SetConfirmIdentityProviderAttributes gets a reference to the given bool and assigns it to the ConfirmIdentityProviderAttributes field.
-func (o *SignOnPolicyActionIDFirstAllOf) SetConfirmIdentityProviderAttributes(v bool) {
-	o.ConfirmIdentityProviderAttributes = &v
 }
 
 // GetDiscoveryRules returns the DiscoveryRules field value if set, zero value otherwise.
@@ -243,9 +205,6 @@ func (o *SignOnPolicyActionIDFirstAllOf) SetSocialProviders(v []SignOnPolicyActi
 
 func (o SignOnPolicyActionIDFirstAllOf) MarshalJSON() ([]byte, error) {
 	toSerialize := map[string]interface{}{}
-	if o.ConfirmIdentityProviderAttributes != nil {
-		toSerialize["confirmIdentityProviderAttributes"] = o.ConfirmIdentityProviderAttributes
-	}
 	if o.DiscoveryRules != nil {
 		toSerialize["discoveryRules"] = o.DiscoveryRules
 	}