+++ management/model_sign_on_policy_action_login.go
@@ -25,8 +25,6 @@ type SignOnPolicyActionLogin struct {
 	Priority int32 `json:"priority"`
 	SignOnPolicy *SignOnPolicyActionCommonSignOnPolicy `json:"signOnPolicy,omitempty"`
 	Type EnumSignOnPolicyType `json:"type"`
-	// A boolean that specifies whether users must confirm data returned from an identity provider prior to registration. Users can modify the data and omit non-required attributes. Modified attributes are added to the user's profile during account creation. This is an optional property. If omitted, the default value is set to false.
-	ConfirmIdentityProviderAttributes *bool `json:"confirmIdentityProviderAttributes,omitempty"`
 	// A boolean that if set to true and if the user's account is locked (the account.canAuthenticate attribute is set to false), then social sign on with an external identity provider is prevented.
 	EnforceLockoutForIdentityProviders *bool `json:"enforceLockoutForIdentityProviders,omitempty"`
 	Recovery *SignOnPolicyActionLoginAllOfRecovery `json:"recovery,omitempty"`
@@ -43,8 +41,6 @@ func NewSignOnPolicyActionLogin(priority int32, type_ EnumSignOnPolicyType) *Sig
 	this := SignOnPolicyActionLogin{}
 	this.Priority = priority
 	this.Type = type_
-	var confirmIdentityProviderAttributes bool = false
-	this.ConfirmIdentityProviderAttributes = &confirmIdentityProviderAttributes
 	return &this
 }
 
@@ -53,8 +49,6 @@ func NewSignOnPolicyActionLogin(priority int32, type_ EnumSignOnPolicyType) *Sig
 // but it doesn't guarantee that properties required by API are set
 func NewSignOnPolicyActionLoginWithDefaults() *SignOnPolicyActionLogin {
 	this := SignOnPolicyActionLogin{}
-	var confirmIdentityProviderAttributes bool = false
-	this.ConfirmIdentityProviderAttributes = &confirmIdentityProviderAttributes
 	return &this
 }
 
@@ -266,38 +260,6 @@ func (o *SignOnPolicyActionLogin) SetType(v EnumSignOnPolicyType) {
 	o.Type = v
 }
 
-// GetConfirmIdentityProviderAttributes returns the ConfirmIdentityProviderAttributes field value if set, zero value otherwise.
-func (o *SignOnPolicyActionLogin) GetConfirmIdentityProviderAttributes() bool {
-	if o == nil || o.ConfirmIdentityProviderAttributes == nil {
-		var ret bool
-		return ret
-	}
-	return *o.ConfirmIdentityProviderAttributes
-}
-
-// GetConfirmIdentityProviderAttributesOk returns a tuple with the ConfirmIdentityProviderAttributes field value if set, nil otherwise
-// and a boolean to check if the value has been set.
-func (o *SignOnPolicyActionLogin) GetConfirmIdentityProviderAttributesOk() (*bool, bool) {
-	if o == nil || o.ConfirmIdentityProviderAttributes == nil {
-		return nil, false
-	}
-	return o.ConfirmIdentityProviderAttributes, true
-}
-
-// HasConfirmIdentityProviderAttributes returns a boolean if a field has been set.
-func (o *SignOnPolicyActionLogin) HasConfirmIdentityProviderAttributes() bool {
-	if o != nil && o.ConfirmIdentityProviderAttributes != nil {
-		return true
-	}
-
-	return false
-}
-
-// SetConfirmIdentityProviderAttributes gets a reference to the given bool and assigns it to the ConfirmIdentityProviderAttributes field.
-func (o *SignOnPolicyActionLogin) SetConfirmIdentityProviderAttributes(v bool) {
-	o.ConfirmIdentityProviderAttributes = &v
-}
-
 // GetEnforceLockoutForIdentityProviders returns the EnforceLockoutForIdentityProviders field value if set, zero value otherwise.
 func (o *SignOnPolicyActionLogin) GetEnforceLockoutForIdentityProviders() bool {
 	if o == nil || o.EnforceLockoutForIdentityProviders == nil {
@@ -448,9 +410,6 @@ func (o SignOnPolicyActionLogin) MarshalJSON() ([]byte, error) {
 	}
 	if true {
 		toSerialize["type"] = o.Type
-	}
-	if o.ConfirmIdentityProviderAttributes != nil {
-		toSerialize["confirmIdentityProviderAttributes"] = o.ConfirmIdentityProviderAttributes
 	}
 	if o.EnforceLockoutForIdentityProviders != nil {
 		toSerialize["enforceLockoutForIdentityProviders"] = o.EnforceLockoutForIdentityProviders