+++ management/CHANGELOG.md
@@ -4,6 +4,7 @@
 * **Bug fix** - Restructure `SignOnPolicyAction` model to properly separate policy action attributes from each other
 * **Bug fix** - Restructure `SignOnPolicyActionMFA` model to correct `noDevicesMode` attribute (fixed incorrect `noDeviceMode`)
 * **Bug fix** - Corrections to `SignOnPolicyActionCommonConditions` model
+* **Bug fix** - Moved `confirmIdentityProviderAttributes` boolean to the registration sub model of `SignOnPolicyAction` 
 
 # v0.4.0 (2022-08-10)